Move model and mommy tests out of all_tests
